+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - [Improvement] Align attributes available in the instrumentation bus for listener related events.
 - [Improvement] Include consumer group id in consumption related events (#1093)
 - [Improvement] Delegate pro components loading to Zeitwerk
+- [Improvement] Include `Datadog::LoggerListener` for tracking logger data with DataDog (@bruno-b-martins)
 - [Refactor] Remove unused logger listener event handler.
 - [Refactor] Internal refactoring of routing validations flow.
 - [Refactor] Reorganize how routing related features are represented internally to simplify features management.